backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.25
langchain-google-genai==2.1.4
langgraph==0.4.1
pydantic==2.11.1
python-dotenv==1.1.0
